Try to solve #12851 숨바꼭질2 (gold V)
There is problem with current solution. Some fastest paths to K are blocked because isVisited is checked as true. For example, for the
first test case in the problem, there is a path from 5 to 17 like 5 -> 10 -> 9 -> 18 -> 17. It takes two attempts to go from 9 to 17. But if there is a case where there is only one attempt left at 9 before the path, like 5 -> 4 -> 8 -> 9 -> ? this attempt will be failed, and isVisited for 9 checked true. Eventually, it blocked the path to go from 9 to 17 the path explained earlier.

So, I have to find another solution.